Unite sorting steps.
