--- src/Processors/QueryPlan/MergingSortedStep.cpp
@@ -1,85 +1,0 @@
-#include <Processors/QueryPlan/MergingSortedStep.h>
-#include <QueryPipeline/QueryPipelineBuilder.h>
-#include <Processors/Merges/MergingSortedTransform.h>
-#include <IO/Operators.h>
-#include <Common/JSONBuilder.h>
-
-namespace DB
-{
-
-static ITransformingStep::Traits getTraits(size_t limit)
-{
-    return ITransformingStep::Traits
-    {
-        {
-            .preserves_distinct_columns = true,
-            .returns_single_stream = true,
-            .preserves_number_of_streams = false,
-            .preserves_sorting = false,
-        },
-        {
-            .preserves_number_of_rows = limit == 0,
-        }
-    };
-}
-
-MergingSortedStep::MergingSortedStep(
-    const DataStream & input_stream,
-    SortDescription sort_description_,
-    size_t max_block_size_,
-    UInt64 limit_)
-    : ITransformingStep(input_stream, input_stream.header, getTraits(limit_))
-    , sort_description(std::move(sort_description_))
-    , max_block_size(max_block_size_)
-    , limit(limit_)
-{
-    /// TODO: check input_stream is partially sorted (each port) by the same description.
-    output_stream->sort_description = sort_description;
-    output_stream->sort_mode = DataStream::SortMode::Stream;
-}
-
-void MergingSortedStep::updateLimit(size_t limit_)
-{
-    if (limit_ && (limit == 0 || limit_ < limit))
-    {
-        limit = limit_;
-        transform_traits.preserves_number_of_rows = false;
-    }
-}
-
-void MergingSortedStep::transformPipeline(QueryPipelineBuilder & pipeline, const BuildQueryPipelineSettings &)
-{
-    /// If there are several streams, then we merge them into one
-    if (pipeline.getNumStreams() > 1)
-    {
-
-        auto transform = std::make_shared<MergingSortedTransform>(
-                pipeline.getHeader(),
-                pipeline.getNumStreams(),
-                sort_description,
-                max_block_size, limit);
-
-        pipeline.addTransform(std::move(transform));
-    }
-}
-
-void MergingSortedStep::describeActions(FormatSettings & settings) const
-{
-    String prefix(settings.offset, ' ');
-    settings.out << prefix << "Sort description: ";
-    dumpSortDescription(sort_description, input_streams.front().header, settings.out);
-    settings.out << '\n';
-
-    if (limit)
-        settings.out << prefix << "Limit " << limit << '\n';
-}
-
-void MergingSortedStep::describeActions(JSONBuilder::JSONMap & map) const
-{
-    map.add("Sort Description", explainSortDescription(sort_description, input_streams.front().header));
-
-    if (limit)
-        map.add("Limit", limit);
-}
-
-}
